// src/modules/auth/dtos/generated.ts
export interface components {
  headers: never;
  parameters: never;
  pathItems: never;
  requestBodies: never;
  responses: never;
  schemas: {
    SignUpRequest: {
      fullName: string;
      passportNumber: string;
      nationality: string;
      /** Format: date */
      dateOfBirth: string;
      /** @enum {string} */
      gender: "M" | "F";
      /** Format: date */
      expiryDate: string;
    } & {
      /** Format: email */
      email: string;
      password: string;
      firstName: string;
      lastName: string;
    };
    SignUpResponse: {
      userId: string;
      /** Format: email */
      email: string;
      firstName?: string;
      lastName?: string;
      requiresOtpVerification: boolean;
      message?: string;
    };
    SignInRequest: {
      /** Format: email */
      email: string;
      password: string;
    };
    SignInResponse: {
      accessToken: string;
      refreshToken?: string;
      user: {
        userId: string;
        /** Format: email */
        email: string;
        firstName?: string;
        lastName?: string;
        /** @enum {string} */
        role?: "user" | "admin";
      };
    };
    VerifyOtpRequest: {
      /** Format: email */
      email: string;
      otp: string;
    };
    VerifyOtpResponse: {
      success: boolean;
      message: string;
      accessToken?: string;
      user?: {
        userId: string;
        /** Format: email */
        email: string;
      };
    };
    LogoutResponse: {
      /** @description status */
      status: string;
    };
  };
}